+++ src/pybot/bot/handlers/__init__.py
@@ -3,3 +3,4 @@ from aiogram import Router, F
 from .common import common_router as common_router
 from .points import points_router as points_router
 from .profile import profile_router as profile_router  #!!! Костыль вывода профиля (Нужно перепроверить и улучшить)
+from .roles import roles_router as roles_router